// src/app/models/cities.model.ts
export interface CityData {
  admin1_code: string;
  admin2_code: string;
  admin3_code: string;
  admin4_code: string;
  alternatenames: Array<string>;
  asciiname: string;
  cc2: string;
  country_code: string;
  dem: string;
  elevation: string;
  feature_class: string;
  feature_code: string;
  geonameid: string;
  latitude: string;
  longitude: string;
  modification_date: string;
  name: string;
  population: string;
  timezone: string;
}